+++ client/web/src/enterprise/app/setup/AppSetupWizard.tsx
@@ -42,9 +42,6 @@ const APP_SETUP_STEPS: StepConfiguration[] = [
         path: 'all-set',
         component: AppAllSetSetupStep,
         nextURL: '/',
-        onView: () => {
-            localStorage.setItem('app.setup.finished', 'true')
-        },
         onNext: async () => {
             await appWindow.setResizable(true)
             await appWindow.setSize(new LogicalSize(1024, 768))
@@ -62,12 +59,7 @@ export const AppSetupWizard: FC<TelemetryProps> = ({ telemetryService }) => {
 
     const handleStepChange = useCallback(
         (nextStep: StepConfiguration): void => {
-            const currentStepIndex = APP_SETUP_STEPS.findIndex(step => step.id === nextStep.id)
-            const isLastStep = currentStepIndex === APP_SETUP_STEPS.length - 1
-
-            // Reset the last visited step if you're on the last step in the
-            // setup pipeline
-            setStepId(!isLastStep ? nextStep.id : '')
+            setStepId(nextStep.id)
         },
         [setStepId]
     )